+++ vlasov/solvers/vlasov/PETScNLVlasovArakawaJ4RK2.pyx
@@ -63,9 +63,9 @@ cdef class PETScVlasovSolver(PETScVlasovSolverBase):
     @cython.boundscheck(False)
     @cython.wraparound(False)
     def jacobian(self, Vec K, Vec Y):
-        cdef np.int64_t a, i, j
-        cdef np.int64_t ix, iy, jx, jy
-        cdef np.int64_t xe, xs, ye, ys
+        cdef int a, i, j
+        cdef int ix, iy, jx, jy
+        cdef int xe, xs, ye, ys
         
         cdef double jpp_J1, jpc_J1, jcp_J1
         cdef double jcc_J2, jpc_J2, jcp_J2
@@ -138,9 +138,9 @@ cdef class PETScVlasovSolver(PETScVlasovSolverBase):
     @cython.boundscheck(False)
     @cython.wraparound(False)
     def function(self, Vec K, Vec Y):
-        cdef np.int64_t a, i, j
-        cdef np.int64_t ix, iy, jx, jy
-        cdef np.int64_t xe, xs, ye, ys
+        cdef int a, i, j
+        cdef int ix, iy, jx, jy
+        cdef int xe, xs, ye, ys
         
         self.Fh.copy(self.Fave)
         self.Fave.axpy(0.5 * self.grid.ht, K)
